# internal/controller/api/api.spec.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "description": "Cloud-Connector API",
      "url": "/api/cloud-connector/"
    }
  ],
  "info": {
    "description": "Cloud-Connector Service",
    "version": "1.0.0",
    "title": "Cloud-Connector Service"
  },
...
  "paths": {
...
    "/v1/message": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "Submit a message to be routed to a customers environment.  This method will take the permitted tenants into consideration when looking up the connection details.  The value of the x-rh-cloud-connector-org-id header is compared against the permitted tenants to locate the connection.",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "PSKAuthClientID": [],
            "PSKAuthAccount": [],
            "PSKAuthOrgID": [],
            "PSKAuthKey": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "No connection to the target connected client"
          }
        }
      }
    },
    "/v1/connection_status": {
      "post": {
        "operationId": "v1.connection.status.multiorg",
        "tags": [
          "api"
        ],
        "summary": "Retrieve the status of the connection.  This method will take the permitted tenants into consideration when looking up the connection details.  The value of the x-rh-cloud-connector-org-id header is compared against the permitted tenants to locate the connection.",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "PSKAuthClientID": [],
            "PSKAuthAccount": [],
            "PSKAuthOrgID": [],
            "PSKAuthKey": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "No connection to the target connected client"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rh-identity",
        "description": "Base64-encoded JSON identity header provided by 3Scale. Contains an account number of the user issuing the request. Format of the JSON: {\"identity\": {\"account_number\": \"12345678\"}}"
      },
      "PSKAuthClientID": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rh-cloud-connector-client-id",
        "description": "Name of the service that is making the request"
      },
      "PSKAuthKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rh-cloud-connector-psk",
        "description": "Key supplied by the platform infrastructure team"
      },
      "PSKAuthAccount": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rh-cloud-connector-account",
        "description": "Account the request is being made on behalf of"
      },
      "PSKAuthOrgID": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rh-cloud-connector-org-id",
        "description": "Org ID the request is being made on behalf of"
      }
    },
    "schemas": {
      "MessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "MessageRequest": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string"
          },
          "recipient": {
            "type": "string"
          },
          "payload": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "directive": {
            "type": "string"
          }
        }
      },
...
      "ConnectionStatusRequest": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string"
          },
          "node_id": {
            "type": "string"
          }
        }
      },
      "ConnectionStatusResponse": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "dispatchers": {
            "type": "object"
          },
          "canonical_facts": {
            "$ref": "#/components/schemas/CanonicalFacts"
          },
          "tags": {
            "$ref": "#/components/schemas/Tags"
          }
        }
      },
...
      "ConnectionStatus": {
        "type": "string",
        "enum": [
          "connected",
          "disconnected"
        ]
      },
      "CanonicalFacts": {
        "type": "object",
        "properties": {
          "insights_id": {
            "type": "string",
            "example": "8d21f37d-bc44-4ade-bdf6-6137f8349af6"
          },
          "machine_id": {
            "type": "string",
            "example": "8d21f37d-bc44-4ade-bdf6-6137f8349af6"
          },
          "bios_uuid": {
            "type": "string",
            "example": "8d21f37d-bc44-4ade-bdf6-6137f8349af6"
          },
          "subscription_manager_id": {
            "type": "string",
            "example": "8d21f37d-bc44-4ade-bdf6-6137f8349af6"
          },
          "ip_addresses": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "127.10.62.12"
            }
          },
          "mac_addresses": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "00:00:5e:00:53:af"
            }
          },
          "fqdn": {
            "type": "string",
            "example": "fredflintstone.bedrock.com"
          }
        }
      },
      "Tags": {
        "type": "object",
        "example": {
          "key1": "value1",
          "key2": "value2"
        }
      }
    }
  }
}